pyproject.toml
[project]
name = "sigflor-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Gabriel Vieira",email = "gabrieldacruzvieira08@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg[binary] (>=3.2.12,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-filter (>=25.2,<26.0)",
]

[tool.poetry]
packages = [{include = "sigflor_server", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
